# backend/docs/swagger.yaml
basePath: /
definitions:
...
  main.Transaction:
    properties:
      amount:
        type: number
      assigned_to:
        items:
          type: string
        type: array
      card_number:
        type: string
      category_id:
        type: string
      created_at:
        type: string
      date_uploaded:
        type: string
      description:
        type: string
      file_name:
        type: string
      id:
        type: string
      posted_date:
        type: string
      transaction_date:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: API for the Joint Analysis expense tracking application
  title: Joint Analysis API
  version: "1.0"
paths:
...
  /api/transactions:
    delete:
      description: Clear all active transactions from the database
      produces:
      - application/json
      responses:
        "200":
          description: All transactions cleared successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete all transactions
      tags:
      - transactions
    get:
      description: Retrieve all active (non-archived) transactions from the database
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            items:
              $ref: '#/definitions/main.Transaction'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all transactions
      tags:
      - transactions
  /api/transactions/{id}:
    delete:
      description: Delete a specific transaction by ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete single transaction
      tags:
      - transactions
  /api/transactions/{id}/assign:
    put:
      consumes:
      - application/json
      description: Assign a specific transaction to one or more people
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Assignment data with array of person names
        in: body
        name: assignment
        required: true
        schema:
          properties:
            assigned_to:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated transaction with assignments
          schema:
            $ref: '#/definitions/main.Transaction'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Assign transaction to person
      tags:
      - transactions
  /api/transactions/{id}/category:
    put:
      consumes:
      - application/json
      description: Update the category assignment for a specific transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Category assignment data
        in: body
        name: category
        required: true
        schema:
          properties:
            category_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated transaction with new category
          schema:
            $ref: '#/definitions/main.Transaction'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update transaction category
      tags:
      - transactions
...
  /api/upload-csv:
    post:
      consumes:
      - multipart/form-data
      description: Upload a CSV file containing transaction data. Returns the successfully
        imported transactions and count of skipped rows.
      parameters:
      - description: CSV file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Upload successful - returns message, transactions array, and
            skipped_rows count
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Upload CSV file
      tags:
      - transactions
swagger: "2.0"